requirements.txt
streamlit
openai
youtube-transcript-api
python-dotenv
watchdog